requirements.txt
streamlit>=1.39.0
pydantic>=2.8.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1